% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{erf_compile}
\alias{erf_compile}
\title{Compile erf data to an R data (rds) object}
\usage{
erf_compile(version_id, format = "rds", folder = NA)
}
\arguments{
\item{version_id}{character temp database version string}

\item{format}{"rds"}

\item{folder}{file.path to data folder.}
}
\description{
Compile erf data from compressed files
}
\examples{
\dontrun{
erf_compile("1", format = "rds")
}
}
