% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{erf_get}
\alias{erf_get}
\title{Get Enhanced River Reach data}
\usage{
erf_get(version_id, skip = NA, dest_folder = NA, compile = TRUE)
}
\arguments{
\item{version_id}{version id}

\item{skip}{numeric vector of lines to skip on file read. optional.}

\item{dest_folder}{file.path optional will default to the location returned by \code{\link[rappdirs]{user_data_dir}}.}

\item{compile}{logical perform on-the-fly compilation to rds?}
}
\description{
Retrieves external files and store in file cache.
}
\details{
Get Enhanced River Reach data.
}
\examples{
\dontrun{
erf_get(version = "1")
}
}
