% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{erf_load}
\alias{erf_load}
\title{erf_load}
\usage{
erf_load(version_id, format = "rds", fpath = NA)
}
\arguments{
\item{version_id}{character database version string}

\item{format}{character choice of rds or dbf}

\item{fpath}{file.path optionally specify custom location of rds file}
}
\description{
Load files from local file system
}
\examples{
\dontrun{
erf  <- erf_load("1")
}
}
