% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux.R
\name{get_hydro_flux}
\alias{get_hydro_flux}
\title{Get Hydro Flux}
\usage{
get_hydro_flux(id_upstream, id_downstream, p_flow)
}
\arguments{
\item{id_upstream}{integer vector}

\item{id_downstream}{integer vector}

\item{p_flow}{data.frame}
}
\value{
data.frame
}
\description{
Get water flux given upstream and downstream nodes.
}
