% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_id}
\alias{get_id}
\title{Get the ERF reachID of a Point}
\usage{
get_id(lon, lat, lines, poly)
}
\arguments{
\item{lon}{numeric longitude}

\item{lat}{numeric lattitude}

\item{lines}{sf polyline}

\item{poly}{sf polygon}
}
\description{
Get the reachID corresponding with a query point-waterbody overlay.
}
