% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux.R
\name{get_p_flux}
\alias{get_p_flux}
\title{Get Phosphorus Flux}
\usage{
get_p_flux(id_upstream, id_focal, id_downstream, p_inc, p_upstream)
}
\arguments{
\item{id_upstream}{integer vector}

\item{id_focal}{integer vector}

\item{id_downstream}{integer vector}

\item{p_inc}{data.frame}

\item{p_upstream}{data.frame}
}
\value{
data.frame
}
\description{
Get phosphorus flux given upstream and downstream nodes as well as
 incremental and total load data.
}
