% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{p_compile}
\alias{p_compile}
\title{Compile phosphorus data to an R data (rds) object}
\usage{
p_compile(version_id, format = "rds", folder = NA, skip = NA)
}
\arguments{
\item{version_id}{character temp database version string}

\item{format}{"rds"}

\item{folder}{file.path to data folder.}

\item{skip}{numeric vector of lines to skip on file read. optional.}
}
\description{
Compile phosphorus erf data from flat files
}
\examples{
\dontrun{
p_compile("1")
}
}
