% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{p_load}
\alias{p_load}
\title{Load phosphorus data}
\usage{
p_load(version_id, format = "rds", fpath = NA)
}
\arguments{
\item{version_id}{character database version string}

\item{format}{character choice of rds}

\item{fpath}{file.path optionally specify custom location of rds file}
}
\description{
Load files from local file system
}
\examples{
\dontrun{
p  <- p_load("1")
}
}
