% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{p_ls}
\alias{p_ls}
\title{p_ls}
\usage{
p_ls(version_id, ...)
}
\arguments{
\item{version_id}{character version id}

\item{...}{extra arguments passed to list.files}
}
\description{
p_ls
}
\examples{
\dontrun{
p_ls("1")
}
}
