% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{erf_compile}
\alias{erf_compile}
\title{Compile data to R data (rds) object}
\usage{
erf_compile(version_id, format = "rds", folder = NA)
}
\arguments{
\item{version_id}{character temp database version string}

\item{format}{character choice of "rds" or "sqlite"}

\item{folder}{file.path to data folder.}
}
\description{
Compile data from component flat files
}
\examples{
\dontrun{
erf_compile("1", format = "rds")
}
}
