% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{erf_ls}
\alias{erf_ls}
\title{erf_ls}
\usage{
erf_ls(version_id, ...)
}
\arguments{
\item{version_id}{character version id}

\item{...}{extra arguments passed to list.files}
}
\description{
erf_ls
}
\examples{
\dontrun{
erf_ls("1")
}
}
