% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network}
\alias{network}
\title{Return sparrow network}
\usage{
network(lon, lat, lines, dbf, polygon)
}
\arguments{
\item{lon}{longitude}

\item{lat}{latitude}

\item{lines}{linestring}

\item{dbf}{dbf object}

\item{polygon}{polygon}
}
\description{
Return sparrow network
}
